FAIR-1779 重模块动态加载 v1
