+++ src/app/components1/topBar/topBar.ts
@@ -7,12 +7,12 @@
  * nextImg:右侧图标路径
  */
 // ================================ 导入
-import { rippleShow } from '../../../chat/client/app/logic/logic';
 import { Json } from '../../../pi/lang/type';
 import { notify } from '../../../pi/widget/event';
 import { Forelet } from '../../../pi/widget/forelet';
 import { Widget } from '../../../pi/widget/widget';
 import { register } from '../../store/memstore';
+import { rippleShow } from '../../utils/tools';
 
 interface Props {
     backImg?:string;
