+++ src/app/components1/walletAssetList/walletAssetList.ts
@@ -1,12 +1,11 @@
 /**
  * wallet home asset list
  */
-import { rippleShow } from '../../../chat/client/app/logic/logic';
 import { notify } from '../../../pi/widget/event';
 import { Forelet } from '../../../pi/widget/forelet';
 import { Widget } from '../../../pi/widget/widget';
 import { getModulConfig } from '../../modulConfig';
-import { getCurrencyUnitSymbol } from '../../utils/tools';
+import { getCurrencyUnitSymbol, rippleShow } from '../../utils/tools';
 // ================================================导出
 // tslint:disable-next-line:no-reserved-keywords
 declare var module: any;
