+++ src/app/core/btc/wallet.ts
@@ -3,6 +3,7 @@
  * BTC HD wallet implementaion
  */
 import { Mnemonic } from '../thirdparty/bip39';
+import { bitcore } from '../thirdparty/bitcore-lib';
 import { BtcApi } from './api';
 
 export type LANGUAGE = 'english' | 'chinese_simplified' | 'chinese_traditional' | 'japanese';
@@ -23,11 +24,6 @@ export class Output {
     public amount: number;
     public chgAddr?: string;
 }
-
-declare var pi_modules;
-const Bitcore = () => {
-    return pi_modules.commonjs.exports.relativeGet('app/core/thirdparty/bitcore-lib').exports.bitcore;
-};
 
 export class BTCWallet {
     // m/bip_number/coin_type/account/internal_or_external/index, we don't use change address
@@ -85,7 +81,7 @@ export class BTCWallet {
             throw new Error('Invalid Mnemonic words!');
         }
         const seed = mn.toSeed(mnemonics, passphrase);
-        const hdpriv = Bitcore().HDPrivateKey.fromSeed(seed, network);
+        const hdpriv = bitcore.HDPrivateKey.fromSeed(seed, network);
 
         const btcwallet = new BTCWallet();
         btcwallet.rootXpriv = hdpriv.toString();
@@ -116,7 +112,7 @@ export class BTCWallet {
         }
         const seed = mn.toSeed(mnemonic, passphrase);
 
-        const hdpriv = Bitcore().HDPrivateKey.fromSeed(seed, network);
+        const hdpriv = bitcore.HDPrivateKey.fromSeed(seed, network);
         const btcwallt = new BTCWallet();
 
         btcwallt.rootXpriv = hdpriv.toString();
@@ -132,7 +128,7 @@ export class BTCWallet {
 
     public static fromSeed(seed: string, network: NETWORK, lang: LANGUAGE): BTCWallet {
         // TODO: check seed ?
-        const hdpriv = Bitcore().HDPrivateKey.fromSeed(seed, network);
+        const hdpriv = bitcore.HDPrivateKey.fromSeed(seed, network);
         const btcwallt = new BTCWallet();
 
         btcwallt.rootXpriv = hdpriv.toString();
@@ -275,8 +271,7 @@ export class BTCWallet {
         console.log('accumlated: ', accumlated);
         console.log('keyset length: ', keySet.length);
 
-        output.amount = Bitcore().Unit.fromBTC(output.amount).toSatoshis();
-        const bitcore = Bitcore();
+        output.amount = bitcore.Unit.fromBTC(output.amount).toSatoshis();
         const rawTx = new bitcore.Transaction().feePerKb(minerFee)
             .from(collected)
             .to(output.toAddr, output.amount)
@@ -338,8 +333,7 @@ export class BTCWallet {
     public async buildRawTransactionFromSingleAddress(address: string, output: Output, minerFee: number): Promise<any> {
         const utxos = await this.coinSelector(address, output.amount * 1e8 + minerFee);
 
-        output.amount = Bitcore().Unit.fromBTC(output.amount).toSatoshis();
-        const bitcore = Bitcore();
+        output.amount = bitcore.Unit.fromBTC(output.amount).toSatoshis();
         const rawTx = new bitcore.Transaction().feePerKb(minerFee)
             .from(utxos)
             .to(output.toAddr, output.amount)
@@ -393,7 +387,7 @@ export class BTCWallet {
         const vout = txInfo.vout;
         const utxos = [];
         let fromAddr = '';
-        const bitcore = Bitcore();
+
         for (let i = 0; i < vin.length; i++) {
             const id = vin[i].txid;
             const vout = vin[i].vout;
@@ -401,9 +395,10 @@ export class BTCWallet {
             fromAddr = address;
             const satoshis = vin[i].valueSat;
 
-            const addr = Bitcore().Address.fromString(address);
-            const script = Bitcore().Script.buildPublicKeyHashOut(addr);
+            const addr = bitcore.Address.fromString(address);
+            const script = bitcore.Script.buildPublicKeyHashOut(addr);
             const scriptPubkey = script.toHex();
+
             const utxo = new bitcore.Transaction.UnspentOutput({
                 txid: id,
                 vout: vout,
@@ -419,7 +414,7 @@ export class BTCWallet {
 
         for (let i = 0; i < vout.length; i++) {
             if (vout[i].scriptPubKey.addresses[0] !== fromAddr) {
-                const value = Bitcore().Unit.fromBTC(vout[i].value).toSatoshis();
+                const value = bitcore.Unit.fromBTC(vout[i].value).toSatoshis();
                 const address = vout[i].scriptPubKey.addresses[0];
                 tx.to(address, value);
             }
@@ -508,7 +503,6 @@ export class BTCWallet {
         }
 
         let path: string;
-        const bitcore = Bitcore();
         const parent = new bitcore.HDPrivateKey(this.rootXpriv);
 
         if (parent.network.name === 'testnet') {
