+++ src/app/logic/dataCenter.ts
@@ -19,7 +19,7 @@ import { fetchLocalTxByHash,fetchTransactionList,getMnemonicByHash } from '../ut
  * 创建事件处理器表
  * @example
  */
-export class DataCenter {
+class DataCenter {
     public timerRef: number = 0;
     public timerRef1: number = 0;
     public updateList: any[] = [];
