+++ src/app/view/wallet/create/home.ts
@@ -7,8 +7,9 @@ import { Forelet } from '../../../../pi/widget/forelet';
 import { Widget } from '../../../../pi/widget/widget';
 import { CreateWalletType } from '../../../logic/localWallet';
 import { getModulConfig } from '../../../modulConfig';
+import { loginSuccess } from '../../../net/login';
 import { deleteAccount, getAllAccount } from '../../../store/memstore';
-import { loginSuccess, popNewLoading, popNewMessage } from '../../../utils/tools';
+import { popNewLoading, popNewMessage } from '../../../utils/tools';
 import { VerifyIdentidy1 } from '../../../utils/walletTools';
 // ============================导出
 // tslint:disable-next-line:no-reserved-keywords
